Check Functions

A collection of functions for checking the validity of function arguments.

Functions:

check_inherits(): Check if an object inherits from a class.

check_installed(): Check if a package is installed.

check_chat(): Check if an object is an ellmer::Chat object.

check_tool(): Check if an object is an ellmer::tool() object.

check_row(): Check if an object is a single-row data frame.

Usage

check_inherits(
  x,
  class,
  x_arg = rlang::caller_arg(x),
  call = rlang::caller_env()
)

check_installed(pkg, arg = rlang::caller_arg(pkg), call = rlang::caller_env())

check_chat(chat, arg = rlang::caller_arg(chat), call = rlang::caller_env())

check_tool(tool, arg = rlang::caller_arg(tool), call = rlang::caller_env())

check_row(row, arg = rlang::caller_arg(row), call = rlang::caller_env())

Arguments

x

An object to check.

class

A character string representing the class to check against.

x_arg

A character string representing the argument name.

call

An environment representing the calling environment.

pkg

A character string representing the package name.

chat

An ellmer::Chat object.

tool

An ellmer::tool() object.

row

A single-row data frame.