Email Address and Domain Server Validation

Validate email addresses and domain servers by checking format and regular expression patterns, MX DNS records, and A DNS records.

Functions:

validate_email_regex(): Validates the format of an email address using a regular expression.

validate_email_mx(): Validates the email address by checking its MX records and falling back to A records if necessary.

validate_domain(): Validates the domain by checking its A and MX records.

Usage

validate_email_regex(email)

validate_email_mx(email, timeout = 5)

validate_domain(domain)

Arguments

email

A character string representing the email address to validate.

timeout

An integer representing the timeout in seconds for DNS queries.

domain

A character string representing the domain to validate.

Value

validate_email_regex(): Returns TRUE if the email address is valid according to the regex pattern, otherwise FALSE.

validate_email_mx(): Returns TRUE if the email address has valid MX records or A records, otherwise FALSE.

validate_domain(): Returns TRUE if the domain has valid A and MX records, otherwise FALSE.

Examples

email <- "jimmy.briggs@noclocks.dev"
validate_email_regex(email)
#> [1] TRUE
validate_email_mx(email)
#> ! No valid MX records found for domain: noclocks.dev. Falling back to A record check.
#> ✖ No valid A records found for domain: noclocks.dev. Email validation failed.
#> [1] FALSE
domain <- sub(".*@", "", email)
validate_domain(domain)
#> ✔ Domain: noclocks.dev has valid A and MX records.
#> [1] TRUE