Determine the Best Match from Multiple Places

This function takes a list of places returned from the Google Maps Places API and compares them to find the best match based on various criteria.

Usage

gmaps_find_best_match(
  places,
  company_name,
  company_address,
  company_phone = NULL
)

Arguments

places

List of places returned from the Google Maps Places API.

company_name

Character string representing the name of the company.

company_address

Character string representing the address of the company.

company_phone

Character string representing the phone number of the company.

Value

A list containing the following elements:

display_name: The display name of the place.

formatted_address: The formatted address of the place.

phone: The phone number of the place.

business_status: The business status of the place.

business_primary_type: The primary type of the business.

business_types: A vector of business types associated with the place.

latitude: The latitude of the place.

longitude: The longitude of the place.

maps_url: The Google Maps URL of the place.

place_id: The unique identifier of the place.

website: The website URL of the place.

domain: The domain extracted from the website URL.

Details

The function evaluates each place based on:

Type of place (e.g., real estate agency, property management company)

String similarity of the company name

String similarity of the address

Phone number area code match

Presence of a website

Business status (operational or not)

The function returns the place with the highest score, indicating the best match.