Geocode Address

Geocode an address using the Google Maps Geocoding API.

Usage

gmaps_geocode_address(address, api_key = get_gmaps_api_key())

Arguments

address

A character string representing the address to geocode.

api_key

A character string representing the Google Maps API key. Will default to the result from get_gmaps_api_key().

Value

A list containing the following elements:

status: A character string indicating the status of the API request.

formatted_address: A character string representing the formatted address.

place_id: A character string representing the place ID.

place_types: A character string representing the place types.

latitude: A numeric value representing the latitude.

longitude: A numeric value representing the longitude.