Google Maps Places Search

Search for a company using the Google Maps Places API.

Usage

gmaps_places_search(
  company_name,
  company_address,
  company_phone = NULL,
  api_key = get_gmaps_api_key()
)

Arguments

company_name

Character string representing the name of the company.

company_address

Character string representing the address of the company.

company_phone

Character string representing the phone number of the company.

api_key

A character string representing the Google Maps API key. Defaults to the result from get_gmaps_api_key().

Value

A list containing the following elements:

website: A character string representing the website URL.

domain: A character string representing the domain extracted from the website URL.

display_name: A character string representing the display name of the place.

formatted_address: A character string representing the formatted address of the place.

phone: A character string representing the phone number of the place.

business_status: A character string representing the business status of the place.

business_type: A character string representing the business type of the place.

latitude: A numeric value representing the latitude of the place.

longitude: A numeric value representing the longitude of the place.

maps_url: A character string representing the Google Maps URL of the place.

place_id: A character string representing the place ID of the place.

Examples

if (FALSE) { # \dontrun{
  gmaps_places_search("Google", "1600 Amphitheatre Parkway, Mountain View, CA")
} # }