Initialize Chat

Initialize a chat session with the OpenAI API using the gpt-4o model (by default) and register the necessary custom function calling tools for performing property investent analysis, Google Maps API functions, and Hunter.io API functions.

Usage

initialize_chat(
  model = "gpt-4o",
  system_prompt = prompt_default(),
  api_key = get_openai_api_key(),
  echo = c("none", "text", "all")
)

Value

An ellmer::chat object configured with the specified system prompt and tools.

See also